requirements.txt
playwright
openai
python-dotenv
pytest
pytest-asyncio
